+++ config/enso/comments.php
@@ -4,6 +4,7 @@ use LaravelEnso\Core\app\Models\UserGroup;
 
 return [
     'editableTimeLimit' => 24 * 60 * 60,
+    'onDelete' => 'cascade',
     'loggableMorph' => [
         'commentable' => [
             UserGroup::class => 'name',