+++ config/enso/discussions.php
@@ -3,6 +3,7 @@
 use LaravelEnso\Core\app\Models\UserGroup;
 
 return [
+    'onDelete' => 'cascade',
     'loggableMorph' => [
         'discussable' => [
             UserGroup::class => 'name',