+++ config/enso/documents.php
@@ -7,6 +7,7 @@ return [
     'linkExpiration' => 60 * 60 * 24,
     'imageWidth' => 2048,
     'imageHeight' => 2048,
+    'onDelete' => 'restrict',
     'loggableMorph' => [
         'documentable' => [
             UserGroup::class => 'name',